app/src/androidTest/java/com/kml/parliamentapp/database/MembersDatabaseTest.kt
package com.kml.parliamentapp.database

import androidx.room.Room
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.test.platform.app.InstrumentationRegistry
import org.junit.Assert.assertEquals
import org.junit.After
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import java.io.IOException


/**
 * This is not meant to be a full set of tests. For simplicity, most of your samples do not
 * include tests. However, when building the Room, it is helpful to make sure it works before
 * adding the UI.
 */

@RunWith(AndroidJUnit4::class)
class MembersDatabaseTest {

    private lateinit var memberDao: MembersDatabaseDao
    private lateinit var db: MembersDatabase

    @Before
    fun createDb() {
        val context = InstrumentationRegistry.getInstrumentation().targetContext
        // Using an in-memory database because the information stored here disappears when the
        // process is killed.
        db = Room.inMemoryDatabaseBuilder(context, MembersDatabase::class.java)
            // Allowing main thread queries, just for testing.
            .allowMainThreadQueries()
            .build()
        memberDao = db.membersDatabaseDao
    }

    @After
    @Throws(IOException::class)
    fun closeDb() {
        db.close()
    }

    @Test
    @Throws(Exception::class)
    fun insertAndGetMember() {
        val member = ParliamentMember(
            1111, 900, "Lauri", "Kenert", "", "sssss"
        )
        memberDao.insert(member)
        val addedMember = memberDao.getById(member.hetekaId)
        assertEquals(addedMember?.hetekaId, member.hetekaId)
    }
}
